Реализован  алгоритм "Решето Эратосфена". Работает на диапазоне от 0 до 100 000 000 примерно 1 минуту.

+++ lab3/task4_var2/task4_var2/PrimeNumbers.h
@@ -2,4 +2,4 @@
 
 #include <set>
 
-std::set<int> GeneratePrimeNumbersSet(int upperBound);
+std::set<int> GeneratePrimeNumbersSet(size_t upperBound);
